include/bits.h: Fir getles64(), getleu64() for 32-bit ints.

+++ include/bits.h
@@ -36,8 +36,8 @@
 #define getleu16(buf, off)      ((uint16_t)((getub((buf), (off)+1) << 8) | getub((buf), (off))))
 #define getles32(buf, off)      ((long)(((int16_t)getles16((buf), (off)+2) << 16) | getleu16((buf), (off))))
 #define getleu32(buf, off)      ((unsigned long)(((unsigned long)getleu16((buf),(off)+2) << 16) | (unsigned long)getleu16((buf), (off))))
-#define getles64(buf, off)      ((long long)((getles32(buf, (off)+4) << 32) | getleu32(buf, (off))))
-#define getleu64(buf, off)      ((unsigned long long)((getleu32(buf, (off)+4) << 32) | getleu32(buf, (off))))
+#define getles64(buf, off)      ((long long)(((long long)getles32(buf, (off)+4) << 32) | getleu32(buf, (off))))
+#define getleu64(buf, off)      ((unsigned long long)(((unsigned long long)getleu32(buf, (off)+4) << 32) | getleu32(buf, (off))))
 extern float getlef32(const char *, int);
 extern double getled64(const char *, int);
